Rename the downstream remote 'gerrithub'
Having a clear differenciation between the branches that will be part of
the upstream repository and the downstream one will allow us to work
upstream and downstream using the same local repository. However, since
gerrit uses 'gerrit' name as the default gerrit repo, we need to rename
the downstream one as 'gerrithub'. The `git review` call will need to
add the option `-r gerrithub` when we want to push downstream.

Change-Id: I8f5ff16efa2111505b227b3ed8efba843badb214