Drop Python2.7 support
This patch removes Python 2.7 dependency due to lack of asyncio support.
Raven needs to watch multiple endpoints and translate them into Neutron
resources. It's better to have the asynchronous manner and handle
multiple endpoints in the single management entity, a daemon or a
service. To achieve it we need the library that helps us to build the
event driven mechanism, but it varies and fragmented in some libraries.
asyncio library introduced in Python 3.4 and it's expected to unify the
foundation for asynchronous I/O mechanism.

This patch would be controversial but OpenStack is shifting toward
supporting Python 3.x in the future release. Kuryr is a young project
where the migration to Python 3.x is easily adopted and it will be run
inside the container eventually. So this limitation shall be accepted
reasonably.

Change-Id: I1fbbc0af125e55d7cca20d41c59587edb076abf6
Signed-off-by: Taku Fukushima <f.tac.mac@gmail.com>