Add Raven service
This patch adds Raven service that takes watchers, watch the
endpoints specified in them and translates their responses.

Raven is initially defined as the service and watch registered endpoints
and call the callbacks associated with the endpoints in its
WATCH_ENDPOINTS_AND_CALLBACKS attribute, which is modified by
register_watchers decorator.

Change-Id: I04facd972f2f4f6d13bb968d18054e828c68d918
Signed-off-by: Taku Fukushima <f.tac.mac@gmail.com>
Signed-off-by: Antoni Segura Puimedon <toni@midokura.com>
Implements: IN-68