Introducing the CNI driver
Add the CNI driver for K8s. The structure proposed tries to be flexible
enough to add an eventual CNI+Rocket.

CNI driver with K8s retrieves the Neutron port and subnets in
"annotations" property of the pod metadata from the K8s API for binding
the created veth endpoint. This data is introduced by Raven watchers.

Signed-off-by: Jaume Devesa <devvesa@gmail.com>
Signed-off-by: Taku Fukushima <f.tac.mac@gmail.com>
Signed-off-by: Antoni Segura Puimedon <toni@midokura.com>

Change-Id: I998a1dfe7739fc6c361a6e8a220e1dfce4bc1dae
Implements: task/70