Update K8sPodsWatcher.translate to create ports
This patch modifies the translate method of K8sPodsWatcher to create
ports when the ADD events of pods are notified to Raven. Raven creates
the port for the newly created pod and add its information required by
the CNI plugin to the metadata of the pod issuing the PATCH request
against the endpoint for it.

Change-Id: Ief432437d8ed7c17ed8c774440ebd4e6043cf8b3
Signed-off-by: Taku Fukushima <f.tac.mac@gmail.com>
Partially-Implements: task/69