Make binding/unbinding suport CNI
In CNI and its K8s implementation, the interface name after it's put
inside the netns, defaults to "eth0", and the path to netns, i.e.,
/proc/1234/ns/net, are given as the environment variables, CNI_IFNAME
and CNI_NETNS.

In the case of Docker, the interface is put inside the netns by Docker.
However, CNI doesn't specify how the interfaces are associated with the
netns, which information is given as the environment variables, and K8s
requires to put the one endpoint of the created veth pair into the netns
for the CNI plugin.

This patch handles the association of the veth endpoint if ifname and
netns are given as the arguments of kuryr.binding.port_bind. In this way
we can share the code for binding interfaces for the CNM plugin and the
CNI plugin.

Change-Id: I8fb3ac767f336be47ed4b4e99a72b2e592112cec
Signed-off-by: Taku Fukushima <f.tac.mac@gmail.com>
Signed-off-by: Antoni Segura Puimedon <toni@midokura.com>
Partially-implements: task/70