Fix reconnect bug when the server sends EOF
When watching for a long time, it is expected that the K8s API server
will EOF the connection to make the watchers reconnect.

This patch makes Raven use a `self._reconnect` which defaults to true
to schedule another watch task when the present one ends with EOF
read from the server.

Scheduling new tasks on the `self._event_loop` was not really
possible when using `run_until_complete` so changing to `run_forever`
was necessary. Of course, this involves doing a bit of housekeeping
with the scheduled tasks. I added some logging to help spot issues
with that.

When reconnecting, it is common for Kubernetes to start the response
with already sent events that, if the reconnection was very fast,
will already have been processed. This patch introduces a capped
size event cache to prevent events from being processed multiple
times.

I also found out that you get much better output from the tests when
something goes wrong if you do

    source .tox/py34/bin/activate
    python -m testtools.run kuryr.tests.unit.test_raven

Signed-off-by: Antoni Segura Puimedon <toni@midokura.com>
Implements task/110

Change-Id: I0adacea4d6ed0e8ca0622e25fcb129318c081ccd