Modify Raven to create the service network and subnet
This patch modifies Raven._ensure_networking_base to create the networks
and the subnets for the services besides the ones for the clusters. They
are connected by the routers, which are supposed to be created for each
namespace but in this patch we assume we only have the single
namespace. Multi namespace support shall come later introducing the
namespace watcher.

To decompose _ensure_networking_base into several pieces,
_get_or_create_service_router, _construct_cluster_network,
_construct_service_network are introduced.

Change-Id: I662ff5da2cc6179129f469badc2a745cc7165917
Signed-off-by: Taku Fukushima <f.tac.mac@gmail.com>