Add pool and VIP deletion to K8sServicesWatcher
This patch adds the deletion logic for the create Neutron Pools and VIPs
when DELETED events of Services are notified. We assume services are
deleted after they're created appropriately. Otherwise
K8sServicesWatcher does nothing because it relies on the annotations
added when the services are created and translated by Raven.

Change-Id: I076eff6a14fe6024c8ac5e2a1a7d08054bd71d87
Signed-off-by: Taku Fukushima <f.tac.mac@gmail.com>
Implements: task/115