Add executor to Raven
This patch modifies Raven to hold an instance of
``concurrent.futures.ThreadPoolExecutor`` as its executor and set it to
the default executor of its event loop. ``delegate`` method of Raven
uses the executor and the passed functions still run concurrently
although the event loop waits for ``run_until_complete``, where
concurrent tasks created ``delegate`` method are finished. This prevents
the unit tests are failed because the cleanup processes are executed
before the tasks created by ``delegate`` method without the executor.

Change-Id: Ie4e78dc241084534ecec0bc273a920d7829dde23
Signed-off-by: Taku Fukushima <f.tac.mac@gmail.com>