Introducing sandbox
This commit introduces the sandbox scripts to be able to deploy all the
services needed by k8s-integration in different containers. It uses the
midonet-sandbox (http://github.com/midonet/midonet-sandbox) to spawn
them.

It is also a good introduction to what the end-to-end tests should use.
So any developer would be able to spawn an environment from where run
these tests.

To test it properly, Docker and Kubectl are required.

Change-Id: I71f1a2c83549d6d8032d89985318dd5962d4f7ce