watcher: Add K8sEndpointsWatcher
This patch adds K8sEndpointsWatcher for the Endpoints translation. The
Endpoints is the resource entity that contains the IP addresses and
ports of the corresponding Services which names are shared in the
metadata properties of the Services and Endpoints'.

K8sEndpointsWatcher translates the subsets of an Endpoint into Neutron
Pool Members, which are associated with the Neutron Pool translated by
K8sServicesWatcher.

With this patch, the deletion of the translated Neutron Pool members
only happens when the service is deleted and it cascades into the
deletion of all of its members. Follow-up patches should take care of
deleting pool members when the replication controller scales down
a service.

Change-Id: I789fc09d594ee14090e441cecc917ebaea5b1a38
Signed-off-by: Taku Fukushima <f.tac.mac@gmail.com>
Signed-off-by: Antoni S. Puimedon <antonisp@celebdor.com>
Partially-Implements: task/115