Add TestK8sWatcherBase.none_future
This patch adds none_future methods as the property of
TestK8sWatcherBase. It caches a created instance of asyncio.Future,
which result is set to the value given as its argument. none_future is
used for any coroutines which don't return the values or return None and
it's tedious to create instances of asyncio.Future every time.

Change-Id: I76fc838f3a5b365cc222ce4ac1b4a334e2b3b12c
Signed-off-by: Taku Fukushima <f.tac.mac@gmail.com>