service_watcher: ignore 'kubernetes' infra service
When running in the sandbox, or when just using kubernetes to deploy
part of the kubernetes components, there will be a service called
'kubernetes' that we do not interact with as of yet.

Taku added code that ignores this service in the endpoints watcher.
This patch does the same for the service watcher.

Implements: task/215
Change-Id: Id35beb2243cb21ddef6402d35419811bb8239338
Signed-off-by: Antoni S. Puimedon <antonisp@celebdor.com>