Introduce the first end-to-end test
This patch introduces an end-to-end test for K8s integration.
It adds common methods in the base test class like:
'assertNeutronPort' and 'assertPingConnection', to simplify the readness
of the test itself.

A k8s python library to do the requests has been added to.

Implements: https://midobugs.atlassian.net/browse/IN-144
Change-Id: Ifa78f729424f604e7fc97ac79393f2711d785639
Signed-off-by: Jaume Devesa <devvesa@gmail.com>