endpoints: watcher to ignore not ready subsets
It is possible for the API server to publish events for Endpoints that
contain subsets that are not yet ready, i.e., that do not have the
'addresses' information.

For now, we'll ignore them. We can in the future consider adding them to
the members so that they are ready in the neutron side even before the
port is up and running, but this is an optimization that I do not think
we should do yet.

Change-Id: Ia34358f33a40528cdbe9298ea2cf3a8262eff9b5
Fixes: task/211
Signed-off-by: Antoni Segura Puimedon <toni@midokura.com>