Make Raven resourceVersion aware
This patch makes Raven to add resourceVersion to the query string of the
endpoint not to receive the outdated events when it recconects. This
prevent Raven from handling outdated and already processed events.
However, this mechanism works only when Raven process is alive and
reconnects by itself. When Raven service is terminated, it starts
receiving events with resourceVersion=0, which means it receives all
events initially.

Change-Id: I2351595a3afd24bf44d5ae99efc4b69e2095e1d9
Signed-off-by: Taku Fukushima <f.tac.mac@gmail.com>
Fixes: task/218