cni: fix kuryr config retrieval
Before this patch, it was using oslo_config.cfg directly, which does
not set a project for oslo to find the configuration directories. By
using kuryr.common.config's init, a project is set and now it searches
for configuration files as it should, for example, now it would find

    /etc/kuryr/kuryr.conf

Thanks to this, now it is possible to set in /etc/kuryr/kuryr.conf:

    [DEFAULT]
    bindir = /opt/kuryr/usr/libexec/kuryr
    [k8s]
    api_root = http:/10.142.0.3:8080

So that we do not need to copy the libexec to dist-packages and cni
can now work with k8s apiservers that reside in different machines.

Change-Id: Ie604b263d49bc728a6c369b1013c9c871c7aab93
Signed-off-by: Antoni Segura Puimedon <toni@midokura.com>