demo: Adapt midonet-agents for fast binding
In the demo we are using the 'latest' midonet/agent containers. These
started using the fast binding mechanism just yesterday. This binding
mechanism relies on writing to a unix domain socket that is served by a
minion in the midonet-agent. For this reason, midonet/agent and
midonet/kubelet have to share access to /var/run. This way they'll find
/var/run/midolman/midolman.sock

Change-Id: I9ae61747d48028a2699434529a67db12004772d9
Signed-off-by: Antoni Segura Puimedon <toni@midokura.com>