testing: stop depending on neutronclient test base
Up until now Kuryr, due to its close relation with neutronclient, it
was reusing neutronclient's cli testcase class as its base. The problem
came from the fact that due to it testing a cli application, it needed
to do a bit of magic with stdout and stdin, which prevents debuggers
from working.

This patch takes only the good things we were getting from
neutronclient's cli testcase class and drops it as a base class, moving
to depend directly on oslotest (as we should have been).

From now on, pdb and ipdb are supported as long as you run the tests
with:

    python -m testtools.run your_test

Change-Id: Idf924f63ace86cd8b850d0a6d2430365752704de
Signed-off-by: Antoni Segura Puimedon <toni@midokura.com>