raven: Add debug log for re-scheduling watchers
Chunked APIs usually terminate connections after a while to make
watchers reconnect as a way to indicate they are still interested in
the content. We had a debug log for when a watcher was finishing for
this reason, but not one for when a new watcher was being scheduled to
resume the resource watch. It could make people think that the watcher
was not running anymore.

Change-Id: Ia2a269afdfa1997b85a87c62e32903f810377f09
Signed-off-by: Antoni Segura Puimedon <toni@midokura.com>